fix: KEEP-611 cast monthly execution count to int so used stays numeric
countMonthlyExecutions sums two COUNT(*) subqueries; postgres-js serializes
bigint as a string, so used came back as a string for the executor billing
guard (the query builder it replaced cast to ::int and returned a number).
Cast each COUNT to ::int so both the Next billing path and the executor get a
numeric used.

+++ lib/billing/execution-limit-core.ts
@@ -42,7 +42,7 @@ export async function countMonthlyExecutions<
   const result = await db.execute<{ count: number }>(
     sql`SELECT
           (
-            SELECT COUNT(*)
+            SELECT COUNT(*)::int
               FROM workflow_executions we
               JOIN workflows w ON we.workflow_id = w.id
              WHERE w.organization_id = ${organizationId}
@@ -51,7 +51,7 @@ export async function countMonthlyExecutions<
           )
           +
           (
-            SELECT COUNT(*)
+            SELECT COUNT(*)::int
               FROM direct_executions de
              WHERE de.organization_id = ${organizationId}
                AND de.created_at >= ${since.toISOString()}
